Another py2 ==> py3 tweak.  The "nogui" version now working in band-hopping mode.
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wspr@4108 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ wspr_nogui.py
@@ -1197,7 +1197,7 @@ def update_nogui():
                 if hopping.coord_bands.get()==1:
                     ns=nsec % 86400
                     ns1=ns % (10*120)
-                    b=ns1/120 + 3
+                    b=int(ns1/120) + 3
                     if b==12: b=2
                     if hopping.hoppingflag[b].get()==0: b=-1
                 if b<0:
